+++ pushswap/pushandswap.c
@@ -40,7 +40,7 @@ int	swapb(t_check *checker)
 
 int	pusha(t_check *checker)
 {
-	int index;
+	long int index;
 
 	index = checker->max_a - 1;
 	if (checker->max_b < 1)
@@ -52,7 +52,7 @@ int	pusha(t_check *checker)
 	}
 	checker->a[0] = checker->b[0];
 	index = 1;
-	while (index < checker->max_b)
+	while ((unsigned)index < checker->max_b)
 	{
 		checker->b[index - 1] = checker->b[index];
 		index++;
@@ -65,7 +65,7 @@ int	pusha(t_check *checker)
 
 int	pushb(t_check *checker)
 {
-	int index;
+	long int index;
 
 	index = checker->max_b;
 	if (checker->max_a < 1)
@@ -77,7 +77,7 @@ int	pushb(t_check *checker)
 	}
 	checker->b[0] = checker->a[0];
 	index = 1;
-	while (index < checker->max_a)
+	while ((unsigned)index < checker->max_a)
 	{
 		checker->a[index - 1] = checker->a[index];
 		index++;
