+++ pushswap/rotate.c
@@ -6,7 +6,7 @@
 /*   By: jcluzet <jo@cluzet.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/21 02:14:44 by jcluzet           #+#    #+#             */
-/*   Updated: 2021/04/21 19:22:43 by jcluzet          ###   ########.fr       */
+/*   Updated: 2021/04/29 03:29:17 by jcluzet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int	rotatea(t_check *checker)
 {
 	int temp;
-	int index;
+	unsigned long int index;
 
 	index = 1;
 	if (checker->max_a < 1)
@@ -34,7 +34,7 @@ int	rotatea(t_check *checker)
 int	reverserotatea(t_check *checker)
 {
 	int temp;
-	int index;
+	unsigned long int index;
 
 	index = checker->max_a - 1;
 	if (checker->max_a < 1)
@@ -53,7 +53,7 @@ int	reverserotatea(t_check *checker)
 int	reverserotateb(t_check *checker)
 {
 	int temp;
-	int index;
+	unsigned long int index;
 
 	index = checker->max_b - 1;
 	if (checker->max_b < 1)
@@ -72,7 +72,7 @@ int	reverserotateb(t_check *checker)
 int	rotateb(t_check *checker)
 {
 	int temp;
-	int index;
+	unsigned long int index;
 
 	index = 0;
 	if (checker->max_b < 1)
